+++ transports/websocket-websys/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.2
+
+- Add support for environments which globally expose `setInterval` and `clearInterval` instead of a window.
+- This crate will now correctly panic if it cannot find the `WebSocket` global JavaScript object.
+
 ## 0.3.1
 
 - Add support for different WASM environments by introducing a `WebContext` that